init: start task 5. migration to ts
