+++ src/resources/users/user.model.ts
