--- src/utils/constants.js
@@ -1,12 +1,0 @@
-const HTTP_STATUS = {
-  OK: 200,
-  CREATED: 201,
-  NO_CONTENT: 204,
-  BAD_REQUEST: 400,
-  NOT_FOUND: 404,
-  INTERNAL_SERVER_ERROR: 500,
-};
-
-module.exports = {
-  HTTP_STATUS,
-};
