+++ src/utils/index.ts
@@ -1,7 +1,11 @@
-const { HTTP_STATUS } = require("./constants");
+// const { HTTP_STATUS } = require("./constants");
+import { HTTP_STATUS } from "./constants";
 const validateId = require("./validateId");
 const findId = require("./findId");
 
+export {
+  HTTP_STATUS
+}
 module.exports = {
   HTTP_STATUS,
   validateId,
